Changes Parser.cs to handle broken item URLs
Accounts for the fact that Amazon no longer seems to prepend its base URL in embedded item links.

Also modifies the code to return null if it can't parse a URL with the regex results. Items with null URLs will be created and added to the grid, but if a user tries to double-click it to go to its Amazon page, it will produce an error dialog ("cannot parse URL").